backend/requirements.txt
flask
flask-cors
opencv-python
numpy
groq
python-dotenv
